requirements.txt
anyio==4.11.0
apsw==3.50.4.0
apswutils==0.1.0
asttokens==3.0.0
babel==2.17.0
beautifulsoup4==4.13.5
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
executing==2.2.1
fastcore==1.8.11
fastlite==0.2.1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
icecream==2.1.8
idna==3.10
itsdangerous==2.2.0
marko==2.2.0
oauthlib==3.3.1
packaging==25.0
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-fasthtml==0.12.29
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.5
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
starlette==0.48.0
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
